requirements.txt
jupyter
selenium
matplotlib
Pillow
scipy

